+++ packages/compiler-cli/src/ngtsc/core/api/index.ts
@@ -1,0 +1,11 @@
+/**
+ * @license
+ * Copyright Google Inc. All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+
+export * from './src/interfaces';
+export * from './src/options';
+export * from './src/public_options';
